requirements.txt
argon2-cffi==20.1.0
astroid==2.4.2
attrs==19.3.0
autopep8==1.5.4
backcall==0.2.0
bleach==3.1.5
cffi==1.14.1
colorama==0.4.3
comtypes==1.1.7
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
future==0.18.2
ipykernel==5.3.4
ipython==7.17.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.17.2
Jinja2==2.11.2
jsonschema==3.2.0
jupyter-client==6.1.6
jupyter-core==4.6.3
kaleido==0.0.2
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.3.0
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.1.0
numexpr==2.7.1
numpy==1.19.1
packaging==20.4
pandas==1.1.0
pandocfilters==1.4.2
parso==0.7.1
pickleshare==0.7.5
Pillow==7.2.0
plotly==4.9.0
prometheus-client==0.8.0
prompt-toolkit==3.0.5
pycodestyle==2.6.0
pycparser==2.20
Pygments==2.6.1
pylint==2.5.3
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
pywin32==228
pywinauto==0.6.8
pywinpty==0.5.7
pyzmq==19.0.2
retrying==1.3.3
rope==0.17.0
schedule==0.6.0
Send2Trash==1.5.0
six==1.15.0
terminado==0.8.3
testpath==0.4.4
toml==0.10.1
tornado==6.0.4
traitlets==4.3.3
typing==3.7.4.1
uiautomation==2.0.5
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xlrd==1.2.0
